Changed type from double to BigDecimal

+++ src/main/java/com/example/banking/service/AccountService.java
@@ -2,6 +2,7 @@ package com.example.banking.service;
 
 import com.example.banking.dto.AccountDto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountService {
@@ -9,9 +10,9 @@ public interface AccountService {
 
     AccountDto getAccountById(Long id);
 
-    AccountDto deposit(Long id, double amount);
+    AccountDto deposit(Long id, BigDecimal amount);
 
-    AccountDto withdraw(Long id, double amount);
+    AccountDto withdraw(Long id, BigDecimal amount);
 
     List<AccountDto> getAllAccounts();
